requirements.txt
anyascii==0.3.1
anyio==3.6.2
arrow==0.15.8
asgiref==3.5.2
astroid==2.12.13
attrs==22.2.0
autohooks==22.11.2
autohooks-plugin-black==22.11.0
autohooks-plugin-isort==22.8.0
autohooks-plugin-pylint==22.8.1
autopep8==1.5.7
beautifulsoup4==4.11.1
black==22.10.0
bleach==4.1.0
boto3==1.26.19
botocore==1.29.19
braintree==3.59.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
colorama==0.4.6
colorful==0.5.5
commonmark==0.9.1
confusable-homoglyphs==3.2.0
crashtest==0.4.1
cryptography==38.0.4
dill==0.3.6
distlib==0.3.6
dj-database-url==0.5.0
Django==4.0.8
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-etc==1.4.0
django-filter==22.1
django-flatpickr==2.0.0
django-hitcount==1.3.5
django-modelcluster==6.0
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-registration==3.3
django-storages==1.13.1
django-taggit==3.1.0
django-timezone-field==5.0
django-treebeard==4.5.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
dulwich==0.20.50
et-xmlfile==1.1.0
filelock==3.8.2
firebase-admin==6.1.0
flake8==3.9.2
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.2
httplib2==0.22.0
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.2.0
isort==5.10.1
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
l18n==2021.3
lazy-object-proxy==1.8.0
lockfile==0.12.2
Markdown==3.4.1
mccabe==0.6.1
more-itertools==9.0.0
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pathspec==0.10.2
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.9.2
platformdirs==2.5.4
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
pontos==22.12.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.7
pyflakes==2.3.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.7
PyMuPDF==1.21.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.14.0
pytz==2022.6
pywin32-ctypes==0.2.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telepath==0.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
trove-classifiers==2022.12.22
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
wagtail==4.2.2
wagtail-color-panel==1.4.0
wagtail-django-recaptcha==1.0
wagtail-markdown==0.10.0
wagtail-metadata==4.0.2
wagtailfontawesome==1.2.1
wagtailmedia==0.11.1
waitress==2.1.2
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4.1
wrapt==1.14.1
zipp==3.11.0
